app/requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.5
incremental==21.3.0
joblib==1.0.1
kombu==5.1.0
lxml==4.6.3
msgpack==1.0.2
multitasking==0.0.9
numpy==1.21.0
pandas==1.2.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5
yfinance==0.1.59
zope.interface==5.4.0
